# 4-bit Integer quantisation