… CMakeLists.txt
… examples/gpt-2/CMakeLists.txt
… examples/gpt-2/README.md
… examples/gpt-2/main.cpp
… examples/gpt-2/quantize.cpp
… examples/gpt-j/CMakeLists.txt
… examples/gpt-j/convert-h5-to-ggml.py
… examples/gpt-j/main.cpp
… examples/gpt-j/quantize.cpp
… examples/utils.cpp
… examples/utils.h
… examples/whisper/CMakeLists.txt
… examples/whisper/convert-pt-to-ggml.py
… examples/whisper/main.cpp
… examples/whisper/quantize.cpp
… examples/whisper/whisper.cpp
… examples/whisper/whisper.h
… tests/test-mul-mat2.c