@ggerganov 2023-03-02
@ocordeiro 

Due to the quantization changes, I had to transpose a few of the tensors in the model.
So this makes the old `ggml` files incompatible with the quantization branch.

In order to make it work, you have to convert the original H5 data using the [convert-h5-to-ggml.py](https://github.com/ggerganov/ggml/blob/master/examples/gpt-j/convert-h5-to-ggml.py) from this branch.
To do that, you need to download the full GPT-J model from here: https://huggingface.co/EleutherAI/gpt-j-6B
And run the command:

```
python3 examples/gpt-j/convert-h5-to-ggml.py ./models/gpt-j-6B 0
```

After you convert the python model to ggml model, you can then use the `gpt-j-quantize` command to quantize the ggml model.

The process is a bit tedious now, but when the implementation is ready, I will upload the quantized models to Hugging Face and it will be easier.