@ocordeiro 2023-03-04
it worked and it's impressive.
here are the results on my M1 Air 8GB:

```
main: mem per token = 15976132 bytes
main:     load time =  2016.22 ms
main:   sample time =    32.71 ms
main:  predict time = 18798.93 ms / 92.61 ms per token
main:    total time = 21609.82 ms
```