@tmzt 2023-03-05
@ocordeiro or anyone else,

can you upload the ggml weights to HF, bittorrent, etc.?