@ocordeiro 2023-03-05
@tmzt it's here until @ggerganov  doesn't launch official version:
https://huggingface.co/ocordeiro/ggml-gpt-j-6b-q4_0