@ggerganov 2023-03-11
@Const-me 
Awesome! Thank you for this.
During the inference the most crucial parts that have to run fast are:

- `quantize_row_q4_0()`:
  https://github.com/ggerganov/ggml/blob/gq/src/ggml.c#L352-L480
- `ggml_vec_dot_q4_0()`: 
  https://github.com/ggerganov/ggml/blob/gq/src/ggml.c#L1172-L1384

For the first one, I have this version, but I don't know if it is optimal yet:

https://github.com/ggerganov/ggml/blob/gq/tests/test-mul-mat2.c#L2038-L2113

For the second one, I have a version for `QK == 64`, but I need one for `QK == 32`:

https://github.com/ggerganov/ggml/blob/gq/tests/test-mul-mat2.c#L1816-L1870

Any advice on the implementation and making it more efficient will be appreciated!