@ggerganov 2023-03-11
@Const-me 

Thank you so much - you are the best!

I just added AVX2 support to `llama.cpp` thanks to your code snippets: https://github.com/ggerganov/llama.cpp/commit/f1eaff4721153a5a5094fd1bd8cbdae7a3c079cc

> Another thing, I don’t understand why are you multiplying two compressed rows? I would expect only the model to be compressed (because using tons of memory, and the compression can be completed offline), but all intermediate tensors be uncompressed FP32 (or at least FP16, upcasting/downcasting vectors is one fast instruction).

Already did that today in the `llama.cpp` repo - it was necessary for consolidating the larger LLaMA models anyway.
Will need to migrate the changes here at some point.

> Generally speaking, I think your CPU matrix multiplication code can be improved by a large factor.

I know! I started doing this with very little knowledge about GEMM and I am sure there is a lot of room for improvements.
Thank you again for all your help.