@Const-me 2023-03-11
@ggerganov About the compression for intermediate tensors, I’ve made another function if you want to try, [dotProduct_q40_f16](https://gist.github.com/Const-me/0c5dc78d5174976eb0779eb6dbb6cde6#file-dotproduct_q40_f16-cpp-L155-L156) I’m not sure what you’ll find, but it’s possible FP16 intermediates might be slightly faster than Q4 compressed.

That block compression is slower than downcasting floats to FP16. And processors often have many megabytes of L3 cache, for example [my processor](https://www.amd.com/en/products/apu/amd-ryzen-7-5700g) has 16MB. The intermediate tensors which were just computed from something else might still be on that cache.