@meakbiyik 2023-03-19
Just to cross reference: 4-bit quantization does not give the expected performance improvement in non-Apple ARM processors. In fact, there is a drastic reduction in performance: https://github.com/ggerganov/whisper.cpp/pull/540#issuecomment-1475167245