@ahoho 2023-04-02
I made a note elsewhere, but I'm finding [q4_1 to be worse than q4_0 in at least one instance.](https://github.com/ggerganov/llama.cpp/issues/397#issuecomment-1493381230)