@ggerganov 2023-04-10
@ahoho 
There might be a bug in the ARM_NEON Q4_1 implementation. I got additional reports indicating that.
Still haven't had time to look into that